changed from solr-5.0.0 to solr

+++ vagrant_solr/solrSetup.sh
@@ -23,20 +23,24 @@ sudo setenforce 0
 ssh-keygen -t rsa -N "" -f ~/.ssh/id_rsa
 
 # Setup Solr
-wget http://apache.spinellicreations.com/lucene/solr/5.0.0/solr-5.0.0.tgz
-#wget http://archive.apache.org/dist/lucene/solr/4.10.4/solr-4.10.4.tgz
-sudo mv solr-5.0.0.tgz /opt/
+if [ -f /vagrant/solr-5.0.0.tgz ]; then
+	cp /vagrant/solr-5.0.0.tgz /opt/
+else
+	wget http://apache.spinellicreations.com/lucene/solr/5.0.0/solr-5.0.0.tgz
+	sudo mv solr-5.0.0.tgz /opt/
+fi
 cd /opt/
 sudo tar zxvf solr-5.0.0.tgz
+mv /opt/solr-5.0.0 /opt/solr
 
-cd /opt/solr-5.0.0/
+cd /opt/solr/
 
 cp /vagrant/solr-in-action.jar server/resources/
 
 cp -r /vagrant/ch15/jts/WEB-INF/ server/webapps/
 cd server/webapps/
 jar -uf solr.war WEB-INF/lib/jts.jar
-cd /opt/solr-5.0.0
+cd /opt/solr/
 rm -r server/webapps/WEB-INF/
 
 bin/solr start -c
